Finished Brand and Item Service Testing, added functionality to check if object exists before calling repository.delete()

+++ API/src/main/java/com/revature/WishListApplication/Service/BrandService.java
@@ -44,7 +44,8 @@ public class BrandService {
     }
 
     public void delete(String id){
-        repository.deleteById(id);
+        if (repository.findById(id).isPresent())
+            repository.deleteById(id);
     }
 
     private BrandDTO BrandToDto(Brand brand){
